config usando mybatis e spring

+++ src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE usuarios (
-                       id BIGINT AUTO_INCREMENT PRIMARY KEY,
-                       name VARCHAR(255),
-                       email VARCHAR(255)
+       id INT AUTO_INCREMENT PRIMARY KEY,
+       name VARCHAR(255),
+       email VARCHAR(255)
 );
